requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.14
kiwisolver==1.4.5
lightning==2.3.2
lightning-utilities==0.11.3.post0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch-lightning==2.3.2
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
sentry-sdk==2.7.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.3.1+cu118
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu118
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.9.0
urllib3==2.2.2
wandb==0.17.4
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
